servers/pyproject.toml
[project]
name = "servers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.8.0",
    "python-dotenv>=1.1.0",
]
